+++ resources/views/admin-home.blade.php
@@ -1,0 +1,3 @@
+<html>
+<p>Welcome to the admin site</p>
+</html>