+++ resources/views/components/layout.blade.php
@@ -25,7 +25,7 @@
         @else
         @if(auth()->user() && auth()->user()->isAdmin())
         <!-- If you are an authenticated user AND the user has an admin role, give them this link. -->
-        <a href="{{ route('dashboard') }}">Admin Dashboard</a>
+        <a href="{{ route('admin.home') }}">Admin Dashboard</a>
         @endif
             <li>
                 <form method="POST" action="{{ route('logout') }}">